Panzer: Refactor interpolation operators
